requirements.txt
openai
fastapi
requests
uvicorn
python-multipart
pandas
python-dotenv
gunicorn 
redis
tiktoken 